split session dependency setup from session logic
